修正若干BUG，让大小和长度更统一

+++ include/vfs/common.h
@@ -8,8 +8,8 @@
 #define VFS_SAFE_FCLOSE(x)		if(x){fclose(x);x=NULL;}
 #define VFS_SAFE_FREE(x)		if(x){free((void*)x);x=NULL;}
 
-#define VFS_CHECK_FREAD(f,b,l)	(f ? (fread((void*)b,1,l,f)        != l ? 0 : 1) : 0 )
-#define VFS_CHECK_FWRITE(f,b,l) (f ? (fwrite((const void*)b,1,l,f) != l ? 0 : 1) : 0 )
+#define VFS_CHECK_FREAD(f,b,l)	(f ? (fread((void*)b,1,(size_t)l,f)        != l ? 0 : 1) : 0 )
+#define VFS_CHECK_FWRITE(f,b,l) (f ? (fwrite((const void*)b,1,(size_t)l,f) != l ? 0 : 1) : 0 )
 
 #define VFS_VERSION				(1)
